move gradient creation to view model
