+++ dressForTheWeather/View Controllers/RecommendationViewController.swift
@@ -40,16 +40,10 @@ final class RecommendationViewController: UIViewController {
     }
     
     private func setBackgroundColor() {
-        // TODO: move into view model
-        guard let mildGreenColor = UIColor(named: "mildGreen"),
-            let warmOrangeColor = UIColor(named: "warmOrange") else { return }
         
-        let gradientLayer = CAGradientLayer()
-        gradientLayer.colors = [mildGreenColor.cgColor, warmOrangeColor.cgColor]
-        gradientLayer.startPoint = CGPoint(x: 0, y: 0)
-        gradientLayer.endPoint = CGPoint(x: 1, y: 1.5)
+        guard let gradientLayer = viewModel.backgroundGradientLayer else { return }
+        
         gradientLayer.frame = view.bounds
-        
         view.layer.insertSublayer(gradientLayer, at: 0)
     }
